Housekeeping Diff
Rationale for the renaming is a stylistic choice.

I like keeping abbreviations capitalized properly when it's convenient to read. e.g. getID(), getURL().
I think that looks a lot nicer than getId() and getUrl()

However, the combination of capital acronyms and camelCase doesn't
always work.

When you have a bunch of capital letters in a row it's hard to see where
the acronym ends and the next word begins.

To compensate for that, acronyms followed by a capital letter should be
lowercase. Acronyms that exist on their own can be all caps.

So things like getID() are fine. But things like getHTTPMethod() will be
getHttpMethod() because it's easier to read.

Other changes here are the addition of some rudimentary authentication
(for now always return true lol), the addition of the Router and
AuthManager abstract classes to encapsulate some convenience functions
for routing.